REFACTOR : state file name, next site map
