+++ lib/storage/index.ts
@@ -2,7 +2,7 @@ type LocalStorageKey = 'access_token' | 'refresh_token' | 'autoComplete'
 
 export class Storage {
 	static getItem(key: LocalStorageKey) {
-		return localStorage.getItem(key)
+		if (typeof window !== 'undefined') return localStorage.getItem(key)
 	}
 
 	static setItem(key: LocalStorageKey, value: string) {
