+++ public/sitemap-0.xml
@@ -1,12 +1,13 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:news="http://www.google.com/schemas/sitemap-news/0.9" xmlns:xhtml="http://www.w3.org/1999/xhtml" xmlns:mobile="http://www.google.com/schemas/sitemap-mobile/1.0" xmlns:image="http://www.google.com/schemas/sitemap-image/1.1" xmlns:video="http://www.google.com/schemas/sitemap-video/1.1">
-<url><loc>http://bumawiki.kro.kr</loc><lastmod>2023-05-11T06:58:26.008Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
-<url><loc>http://bumawiki.kro.kr/accident</loc><lastmod>2023-05-11T06:58:26.008Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
-<url><loc>http://bumawiki.kro.kr/club</loc><lastmod>2023-05-11T06:58:26.008Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
-<url><loc>http://bumawiki.kro.kr/create</loc><lastmod>2023-05-11T06:58:26.008Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
-<url><loc>http://bumawiki.kro.kr/frame</loc><lastmod>2023-05-11T06:58:26.008Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
-<url><loc>http://bumawiki.kro.kr/mypage</loc><lastmod>2023-05-11T06:58:26.008Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
-<url><loc>http://bumawiki.kro.kr/oauth</loc><lastmod>2023-05-11T06:58:26.008Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
-<url><loc>http://bumawiki.kro.kr/student</loc><lastmod>2023-05-11T06:58:26.008Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
-<url><loc>http://bumawiki.kro.kr/teacher</loc><lastmod>2023-05-11T06:58:26.008Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki/accident</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki/club</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki/create</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki/frame</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki/mypage</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki/oauth</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki/popular</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki/student</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
+<url><loc>https://buma.wiki/teacher</loc><lastmod>2023-06-19T06:59:57.121Z</lastmod><changefreq>daily</changefreq><priority>1</priority></url>
 </urlset>
