+++ public/sitemap.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <sitemapindex xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
-<sitemap><loc>http://bumawiki.kro.kr/sitemap-0.xml</loc></sitemap>
+<sitemap><loc>https://buma.wiki/sitemap-0.xml</loc></sitemap>
 </sitemapindex>
